backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
sentence-transformers
openai
anthropic
mistralai
google-generativeai
groq
deepseek
apscheduler
